Got started on extractors. Still have a lot of work to do, as well as implementing them. See autocadExcellParser static parse
