+++ src/main/java/autocadDrawingChecker/autocadData/AutoCADElementMatcher.java
@@ -1,5 +1,6 @@
 package autocadDrawingChecker.autocadData;
 
+import autocadDrawingChecker.autocadData.elements.AutoCADElement;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.function.BiFunction;
