+++ src/main/java/autocadDrawingChecker/autocadData/AutoCADExport.java
@@ -1,5 +1,6 @@
 package autocadDrawingChecker.autocadData;
 
+import autocadDrawingChecker.autocadData.elements.AutoCADElement;
 import java.util.HashMap;
 import java.util.LinkedList;
