+++ src/main/java/autocadDrawingChecker/autocadData/elements/AutoCADElement.java
@@ -1,4 +1,4 @@
-package autocadDrawingChecker.autocadData;
+package autocadDrawingChecker.autocadData.elements;
 
 /**
  * An AutoCADElement represents
