+++ src/main/java/autocadDrawingChecker/grading/implementations/LineEnd.java
@@ -1,7 +1,7 @@
 package autocadDrawingChecker.grading.implementations;
 
 import autocadDrawingChecker.autocadData.AutoCADLine;
-import autocadDrawingChecker.autocadData.AutoCADElement;
+import autocadDrawingChecker.autocadData.elements.AutoCADElement;
 
 /**
  *
