+++ src/main/java/autocadDrawingChecker/start/Main.java
@@ -14,12 +14,11 @@ public class Main {
      */
     public static void main(String[] args) {
         Application app = Application.getInstance();
-        /*
+        
         app.getData()
             .setInstructorFilePath("C:\\Users\\Matt\\Desktop\\AutoCAD Drawing Checker\\sample files to work with\\Check Sample - Master File.xls.xlsx")
-            .setStudentFilePaths("C:\\Users\\Matt\\Desktop\\AutoCAD Drawing Checker\\sample files to work with")
-            .setCriteriaSelected(new CheckDims(), false);
-        */
-        app.createGui();
+            .setStudentFilePaths("C:\\Users\\Matt\\Desktop\\AutoCAD Drawing Checker\\sample files to work with");
+        System.out.println(app.grade().toString());
+        //app.createGui();
     }
 }
